Changing the version to long, as we depend on the last modified date in the PT for doing atomic conflict resolution.
git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/trunk@1761 6771e1e2-0f46-0410-b5d7-d846ff89f743

+++ src/core/com/collabnet/ccf/core/AbstractWriter.java
@@ -222,9 +222,9 @@ public abstract class AbstractWriter<T> extends LifecycleComponent implements
 	 * @return true if update can happen, false if update must not happen and
 	 *         modified generic artifact has to be returned
 	 */
-	public static boolean handleConflicts(int artifactCurrentVersion,
+	public static boolean handleConflicts(long artifactCurrentVersion,
 			GenericArtifact gaDocument) {
-		int lastSyncVersion = -1;
+		long lastSyncVersion = -1;
 		String targetArtifactId = gaDocument.getTargetArtifactId();
 		String lastSyncVersionStr = gaDocument.getTargetArtifactVersion();
 		if (lastSyncVersionStr == null
@@ -233,7 +233,7 @@ public abstract class AbstractWriter<T> extends LifecycleComponent implements
 			lastSyncVersionStr = GenericArtifactHelper.ARTIFACT_VERSION_FORCE_RESYNC;
 		}
 		try {
-			lastSyncVersion = Integer.parseInt(lastSyncVersionStr);
+			lastSyncVersion = Long.parseLong(lastSyncVersionStr);
 		} catch (NumberFormatException e) {
 			String message = "Last successful synchronization version of artifact "
 					+ targetArtifactId
